requirements.txt
websockets
websocket-client
requests
py_clob_client
pandas
matplotlib
web3
scikit-learn